Revert "Revert "Update Sample App Edge View""
This reverts commit 9b8a3ad100ec0fab560412ed020cb7608ac0d116.
