+++ example/ios/Runner/AppDelegate.m
@@ -1,3 +1,14 @@
+/*
+Copyright 2022 Adobe. All rights reserved.
+This file is licensed to you under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License. You may obtain a copy
+of the License at http://www.apache.org/licenses/LICENSE-2.0
+Unless required by applicable law or agreed to in writing, software distributed under
+the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+OF ANY KIND, either express or implied. See the License for the specific language
+governing permissions and limitations under the License.
+*/
+
 #import "AppDelegate.h"
 #import "GeneratedPluginRegistrant.h"
 
@@ -16,10 +27,18 @@
     [AEPMobileCore setLogLevel:AEPLogLevelTrace];
     [AEPMobileCore setPrivacyStatus:AEPPrivacyStatusOptedIn];
     [AEPMobileCore setWrapperType:AEPWrapperTypeFlutter];
+
+     // TODO: Set up the preferred Environment File ID from your mobile property configured in Data Collection UI
+    NSString* ENVIRONMENT_FILE_ID = @"YOUR-APP-ID";
     
     const UIApplicationState appState = application.applicationState;
-    
-    NSArray *extensionsToRegister = @[AEPMobileIdentity.class, AEPMobileLifecycle.class, AEPMobileSignal.class, AEPMobileAssurance.class, AEPMobileEdge.class, AEPMobileEdgeIdentity.class];
+
+    NSArray *extensionsToRegister = @[AEPMobileIdentity.class, 
+                                      AEPMobileLifecycle.class, 
+                                      AEPMobileSignal.class, 
+                                      AEPMobileAssurance.class, 
+                                      AEPMobileEdge.class,
+                                      AEPMobileEdgeIdentity.class];
     
     [AEPMobileCore registerExtensions:extensionsToRegister completion:^{
         if (appState != UIApplicationStateBackground) {
@@ -27,7 +46,7 @@
         }
     }];
     
-    [AEPMobileCore configureWithAppId:@"YOUR-APP-ID"];
+    [AEPMobileCore configureWithAppId: ENVIRONMENT_FILE_ID];
 }
 
 - (void)applicationDidEnterBackground:(UIApplication *)application {
